Replaced QR Demo and Added Combined Project Code
The QR Example code I added was the wrong folder, so I updated it to be
the correct project folder. I also created a separate folder for the QR
code integrated with the interface layout (ScavengerHunt). I need to
look into "Fragments" as a way to get the QR code interactions in a
separate window (rather than using multiple Activities), so everything
is still on the same screen.